Fix printf/scanf-style format strings and cast arguments where needed.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@1487 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ fisher.c
@@ -104,8 +104,8 @@ double fis_get_spamicity(size_t robn, FLOAT P, FLOAT Q )
 
 void fis_print_summary(void)
 {
-    (void)fprintf(stdout, "%-*s %5d %9.2e %9.2e %9.2e %9.2e %9.2e %4.2f\n",
-		  MAXTOKENLEN+2, "N_P_Q_S_s_x_md", fis_stats.robn, 
+    (void)fprintf(stdout, "%-*s %5lu %9.2e %9.2e %9.2e %9.2e %9.2e %4.2f\n",
+		  MAXTOKENLEN+2, "N_P_Q_S_s_x_md", (unsigned long)fis_stats.robn, 
 		  fis_stats.p_pr, fis_stats.q_pr, fis_stats.s.spamicity, robs, robx, min_dev);
 }
